pronunciationChecker.py
For use in Training mode of program. Allows user to listen to words supported by the system which are contained within a corpus file.
Contains compare method which compares the roman script output from the processor to the actual script of the word being trained. Then notifies the user if correct or not, and why.