interactions.py
Without the development of a UI at this stage, the system must be ran using the command prompt. This file contains all the necessary menus, and interactions between the user and the other components of the system.
Isolated to enable simple integration of the other components into a system with a complete GUI.